Fix errors after rebase : move localFarmSubmitter module & remove useless functions
